refactoring task32
