+++ v.zyablov1/task32/client.c
@@ -1,3 +1,4 @@
+#include "client.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -7,9 +8,6 @@
 #include <fcntl.h>
 #include <sys/select.h>
 
-// volatile int exit_flag_custom = 0;  // Глобальная переменная для завершения работы
-#define SOCKET_PATH "/tmp/unix_socket"
-
 int create_nonblocking_socket() {
     int sock = socket(AF_UNIX, SOCK_STREAM, 0);
     if (sock < 0) {
@@ -17,7 +15,6 @@ int create_nonblocking_socket() {
         exit(1);
     }
 
-    // Устанавливаем сокет в неблокирующий режим
     int flags = fcntl(sock, F_GETFL, 0);
     if (flags == -1) {
         perror("fcntl(F_GETFL) failed");
@@ -32,33 +29,31 @@ int create_nonblocking_socket() {
 }
 
 void connect_to_server(int client_sock, const char *socket_path) {
-        // Заполнение и создание структуры адреса сервера
-	struct sockaddr_un server_addr;
-	memset(&server_addr, 0, sizeof(server_addr));
-	server_addr.sun_family = AF_UNIX;
-	strncpy(server_addr.sun_path, socket_path, sizeof(server_addr.sun_path) - 1);
+    struct sockaddr_un server_addr;
+    memset(&server_addr, 0, sizeof(server_addr));
+    server_addr.sun_family = AF_UNIX;
+    strncpy(server_addr.sun_path, socket_path, sizeof(server_addr.sun_path) - 1);
 
-	printf("Connecting..."); fflush(stdout);
+    printf("Connecting..."); fflush(stdout);
 
-	if (connect(client_sock, (struct sockaddr *)&server_addr, sizeof(server_addr)) < 0) {
-		perror("Failed to connect to server");
-		close(client_sock);
-		exit(1);
-	}
+    if (connect(client_sock, (struct sockaddr *)&server_addr, sizeof(server_addr)) < 0) {
+        perror("Failed to connect to server");
+        close(client_sock);
+        exit(1);
+    }
 
-	printf(".\tConnected!\n");
+    printf(".\tConnected!\n");
 }
 
-// Обёртка для отправки сообщения
 void send_message(int client_sock, const char *message) {
-	ssize_t sent_len = send(client_sock, message, strlen(message), 0);
-	if (sent_len < 0) {
-		perror("Failed to send message");
-		close(client_sock);
-		exit(1);
-	}
+    ssize_t sent_len = send(client_sock, message, strlen(message), 0);
+    if (sent_len < 0) {
+        perror("Failed to send message");
+        close(client_sock);
+        exit(1);
+    }
 }
-// 1. Обработка сокета (пришли данные с сервера):
+
 void handle_socket_activity(int client_sock) {
     char buffer[256];
     ssize_t bytes_received = recv(client_sock, buffer, sizeof(buffer), 0);
@@ -73,26 +68,17 @@ void handle_socket_activity(int client_sock) {
         exit(1);
     }
 }
-// 2. Обработка ввода с клавиатуры и отправка сообщения серверу:
+
 void handle_input_activity(int client_sock) {
     char buffer[256];
     printf("Enter message: ");
     if (fgets(buffer, sizeof(buffer), stdin)) {
-        buffer[strcspn(buffer, "\n")] = '\0';  // Убираем символ новой строки
-
-        // Если введено "q", завершаем программу
-        /*
-        if (strcmp(buffer, "q") == 0) {
-            printf("Exiting...\n");
-            exit_flag_custom = 1;
-        }*/
-        // Отправляем введённое сообщение на сервер
+        buffer[strcspn(buffer, "\n")] = '\0';
         if (send(client_sock, buffer, strlen(buffer), 0) < 0) {
             perror("send() failed");
         }
     }
 }
-
 
 void async_client_main_loop(int client_sock) {
     fd_set read_fds;
@@ -100,10 +86,10 @@ void async_client_main_loop(int client_sock) {
 
     while (1) {
         FD_ZERO(&read_fds);
-        FD_SET(client_sock, &read_fds);  // Следим за клиентским сокетом
-        FD_SET(STDIN_FILENO, &read_fds); // Следим за вводом с клавиатуры
+        FD_SET(client_sock, &read_fds);
+        FD_SET(STDIN_FILENO, &read_fds);
 
-        timeout.tv_sec = 5;  // Тайм-аут 5 секунд
+        timeout.tv_sec = 5;
         timeout.tv_usec = 0;
 
         int activity = select(client_sock + 1, &read_fds, NULL, NULL, &timeout);
@@ -115,18 +101,15 @@ void async_client_main_loop(int client_sock) {
             continue;
         }
 
-        // Если сокет клиента готов для чтения
         if (FD_ISSET(client_sock, &read_fds)) {
             handle_socket_activity(client_sock);
         }
 
-        // Если стандартный ввод готов
         if (FD_ISSET(STDIN_FILENO, &read_fds)) {
             handle_input_activity(client_sock);
         }
     }
 }
-
 
 void async_auto_client_main_loop(int client_sock) {
     fd_set read_fds;
@@ -135,9 +118,9 @@ void async_auto_client_main_loop(int client_sock) {
 
     while (1) {
         FD_ZERO(&read_fds);
-        FD_SET(client_sock, &read_fds);  // Следим за клиентским сокетом
+        FD_SET(client_sock, &read_fds);
 
-        timeout.tv_sec = 1;  // Тайм-аут 1 секунда
+        timeout.tv_sec = 1;
         timeout.tv_usec = 0;
 
         int activity = select(client_sock + 1, &read_fds, NULL, NULL, &timeout);
@@ -145,7 +128,6 @@ void async_auto_client_main_loop(int client_sock) {
             perror("select() failed");
             break;
         } else if (activity == 0) {
-            // Если нет активности за 1 секунду, отправляем сообщение
             printf("Auto Client: %s\n", message);
             if (send(client_sock, message, strlen(message), 0) < 0) {
                 perror("send() failed");
@@ -154,13 +136,11 @@ void async_auto_client_main_loop(int client_sock) {
             continue;
         }
 
-        // Если сокет клиента готов для чтения
         if (FD_ISSET(client_sock, &read_fds)) {
             handle_socket_activity(client_sock);
         }
     }
 }
-
 
 int main(int argc, char *argv[]) {
     // Проверка аргументов
@@ -179,4 +159,3 @@ int main(int argc, char *argv[]) {
     close(client_sock);
     return 0;
 }
-
